requirements.txt
discord.py
boto3[crt]
